requirements.txt
requests
python-dotenv
openai
pyyaml